#18 refactor: DummyEntity와 DummyDocument 클래스 추가 및 이동
- DummyEntity 클래스를 health 도메인에 추가
- TestDocument를 DummyDocument로 이름 변경 및 health 도메인으로 이동
